Generic BRM model with fitting defaults set

A generic model used to define default brms settings to be used across all models investigated.

Base_Model(Predictors = NULL, ModelFamily, Data = NULL,
  SetIter = NULL, SetBurn = NULL, SetCores = NULL, Chains = 4,
  SetThin = 1, Silent = TRUE, SaveModel = TRUE, AdaptDelta = NULL,
  MaxTree = NULL)

Arguments

Predictors
A character vector of variables to use as features

ModelFamily
A character string containing the model family to fit.

Data
A dataframe of the data to fit to.

SetIter
Numeric, define the number of iterations

SetBurn
Numeric, define the number of burn in iterations

SetCores
Numeric, set the number of cores to use.

Chains
Numeric, set the number of chains to use for MCMC

SetThin
Nuemeric, set the number of iterations to thin over.

Silent
Logical, defaults to TRUE. Should the results be returned without progress messages.

SaveModel
Logical, defaults to TRUE. Should the model output be saved

AdaptDelta
Numeric, adaptive delta to use

MaxTree
Numeric, the maximum number of trees.

Value

A brm fitted model.

See also

Examples